hal: sdk-ng: sync sdk usb to hal nxp

+++ mcux/mcux-sdk-ng/middleware/usb/utility/usb_eth_mcx_enet_adapter.c
@@ -300,7 +300,7 @@ eth_adapter_err_t ETH_ADAPTER_SendFrame(eth_adapter_frame_buf_t *buffer)
 {
     enet_buffer_struct_t txBuff = {
         .buffer = buffer->payload,
-        .length = buffer->len,
+        .length = (uint16_t)buffer->len,
     };
 
     enet_tx_frame_struct_t txFrame = {
